+++ EnhancedSpectator/Patches/PlayerControllerBPatches.cs
@@ -9,6 +9,7 @@ namespace EnhancedSpectator.Patches
     internal class PlayerControllerBPatches
     {
         internal static bool NightVisionToggled = false;
+        internal static bool WasToggledOnce = false;
 
         /// <summary>
         /// PlayerControllerB patches to enable the use of night vision while spectating
@@ -22,18 +23,34 @@ namespace EnhancedSpectator.Patches
                 PlayerControllerB CurrentPlayer = GameNetworkManager.Instance.localPlayerController;
                 if (CurrentPlayer.isPlayerDead && !StartOfRound.Instance.shipIsLeaving)
                 {
+#if DEBUG
+                    LCES.Log.LogInfo("PlayerControllerBPatches.Prefix => Player Dead");
+#endif
                     NightVisionToggle(__instance, NightVisionToggled);
+                    if (!WasToggledOnce) { WasToggledOnce = true; }
                 }
                 else
                 {
-                    NightVisionToggled = false;
-                    NightVisionToggle(__instance, false);
+                    if (WasToggledOnce)
+                    {
+#if DEBUG
+                        LCES.Log.LogInfo("PlayerControllerBPatches.Prefix => Player Alive");
+#endif
+                        NightVisionToggled = false;
+                        NightVisionToggle(__instance, false);
+                    }
                 }
             }
             else
             {
-                NightVisionToggled = false;
-                NightVisionToggle(__instance, false);
+                if (WasToggledOnce)
+                {
+#if DEBUG
+                    LCES.Log.LogInfo("PlayerControllerBPatches.Prefix => NOT GameNetworkManager");
+#endif
+                    NightVisionToggled = false;
+                    NightVisionToggle(__instance, false);
+                }
             }
         }
 
@@ -44,7 +61,13 @@ namespace EnhancedSpectator.Patches
         /// <param name="ToggleAction"></param>
         private static void NightVisionToggle(PlayerControllerB __instance, bool ToggleAction)
         {
+#if DEBUG
+            LCES.Log.LogInfo(string.Format("PlayerControllerBPatches.NightVisionToggle => ToggleAction: {0}", ToggleAction.ToString()));
+#endif
             if (!ConfigSettings.NightVisionAllowed.Value) { ToggleAction = false; }
+#if DEBUG
+            LCES.Log.LogInfo(string.Format("PlayerControllerBPatches.NightVisionToggle => Allowed => ToggleAction: {0}", ToggleAction.ToString()));
+#endif
             if (ToggleAction)
             {
                 __instance.nightVision.intensity = ConfigSettings.NightVisionIntensity.Value;
@@ -55,7 +78,11 @@ namespace EnhancedSpectator.Patches
             }
             else
             {
-                __instance.nightVision.intensity = 366.9317f;
+                float DarknessModifier = ConfigSettings.ModifyDarkness.Value ? Mathf.Clamp(1f - ConfigSettings.DarknessModifier.Value, 0f, 1f) : 1f;
+#if DEBUG
+                LCES.Log.LogInfo(string.Format("DarknessModifier: {0}", DarknessModifier.ToString()));
+#endif
+                __instance.nightVision.intensity = 366.9317f * DarknessModifier;
                 __instance.nightVision.range = 12f;
                 __instance.nightVision.shadowStrength = 1f;
                 __instance.nightVision.shadows = (LightShadows)0;