+++ EnhancedSpectator/Plugin.cs
@@ -20,7 +20,7 @@ namespace EnhancedSpectator
 
         private const string PluginGUID = "PXC.EnhancedSpectator";
         private const string PluginName = "EnhancedSpectator";
-        private const string VersionString = "1.0.0";
+        private const string VersionString = "1.0.3";
         public static string PluginFullName { get => string.Format("{0} v{1}", PluginName, VersionString); }
         public static string PluginAuthor { get => "PXC"; }
 
@@ -92,11 +92,22 @@ namespace EnhancedSpectator
         /// </summary>
         private void ToggleSpecNightVision()
         {
+#if DEBUG
+            Log.LogInfo("ToggleSpecNightVision");
+#endif
             if (GameNetworkManager.Instance != null && GameNetworkManager.Instance.localPlayerController != null)
             {
+#if DEBUG
+                Log.LogInfo("ToggleSpecNightVision => GameNetworkManager");
+#endif
                 PlayerControllerB CurrentPlayer = GameNetworkManager.Instance.localPlayerController;
                 if (HUDManager.Instance != null && CurrentPlayer.IsOwner && CurrentPlayer.isPlayerDead && !StartOfRound.Instance.shipIsLeaving && (!CurrentPlayer.IsServer || CurrentPlayer.isHostPlayerObject))
                 {
+#if DEBUG
+                    Log.LogInfo(string.Format("ToggleSpecNightVision => GameNetworkManager => HUDManager/CurrentPlayer => NightVisionToggled: {0}", PlayerControllerBPatches.NightVisionToggled));
+                    Log.LogInfo(string.Format("ToggleSpecNightVision => GameNetworkManager => HUDManager/CurrentPlayer => NightVisionAllowed: {0}", ConfigSettings.NightVisionAllowed.Value));
+                    Log.LogInfo(string.Format("ToggleSpecNightVision => GameNetworkManager => HUDManager/CurrentPlayer => NightvisionStatus: {0}", HUDManagerPatches.NightvisionStatus));
+#endif
                     if (ConfigSettings.NightVisionAllowed.Value)
                     {
                         PlayerControllerBPatches.NightVisionToggled = !PlayerControllerBPatches.NightVisionToggled;
@@ -107,11 +118,14 @@ namespace EnhancedSpectator
                     }
 
                     HUDManagerPatches.NightvisionStatus = PlayerControllerBPatches.NightVisionToggled;
+#if DEBUG
+                    Log.LogInfo(string.Format("ToggleSpecNightVision => GameNetworkManager => HUDManager/CurrentPlayer => NightvisionStatus: {0}", HUDManagerPatches.NightvisionStatus));
+#endif
                 }
             }
         }
 
-        #endregion
+#endregion
 
         #region Events
 
@@ -121,6 +135,9 @@ namespace EnhancedSpectator
         /// <param name="context"></param>
         private void OnFlashlightPressed(InputAction.CallbackContext context)
         {
+#if DEBUG
+            Log.LogInfo("OnFlashlightPressed");
+#endif
             if (!context.performed) { return; }
             ToggleSpecFlashlight();
         }
@@ -131,6 +148,9 @@ namespace EnhancedSpectator
         /// <param name="context"></param>
         private void OnNightVisionPressed(InputAction.CallbackContext context)
         {
+#if DEBUG
+            Log.LogInfo("OnNightVisionPressed");
+#endif
             if (!context.performed) { return; }
             ToggleSpecNightVision();
         }