+++ EnhancedSpectator/Utils/ConfigSettings.cs
@@ -7,8 +7,10 @@ namespace EnhancedSpectator.Utils
 {
     public class ConfigSettings
     {
-        public static ConfigEntry<float> NightVisionIntensity;
         public static ConfigEntry<bool> NightVisionAllowed;
+        public static ConfigEntry<float> NightVisionIntensity;
+        public static ConfigEntry<bool> ModifyDarkness;
+        public static ConfigEntry<float> DarknessModifier;
         public static ConfigEntry<string> NightVisionKeybind;
         public static ConfigEntry<bool> FlashlightAllowed;
         public static ConfigEntry<string> FlashlightKeybind;
@@ -26,11 +28,17 @@ namespace EnhancedSpectator.Utils
 
             #region Night Vision
 
+            NightVisionAllowed = config.Bind<bool>("Night Vision", "Allowed", true, "Allow night vision while spectating.");
+            LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(NightVisionAllowed, false));
+
             NightVisionIntensity = config.Bind<float>("Night Vision", "Intensity", 7500f, "This is how bright the night vision makes the environment when enabled.");
             LethalConfigManager.AddConfigItem(new FloatSliderConfigItem(NightVisionIntensity, new FloatSliderOptions { Min = 100f, Max = 100000f, RequiresRestart = false }));
 
-            NightVisionAllowed = config.Bind<bool>("Night Vision", "Allowed", true, "Allow night vision while spectating.");
-            LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(NightVisionAllowed, false));
+            ModifyDarkness = config.Bind<bool>("Night Vision", "Modify Darkness", false, "Some mods (Diversity) change the default darkness intensity value. This setting enables the below option.");
+            LethalConfigManager.AddConfigItem(new BoolCheckBoxConfigItem(ModifyDarkness, false));
+
+            DarknessModifier = config.Bind<float>("Night Vision", "Darkness Modifier", 1f, "This option modifies the default darkness intensity value.");
+            LethalConfigManager.AddConfigItem(new FloatSliderConfigItem(DarknessModifier, new FloatSliderOptions { Min = 0f, Max = 1f, RequiresRestart = false }));
 
             NightVisionKeybind = config.Bind<string>("Night Vision", "Keybind", "<Keyboard>/n", "Input binding to toggle night vision while spectating.");
             LethalConfigManager.AddConfigItem(new TextInputFieldConfigItem(NightVisionKeybind, true));